+++ frontend/src/app/features/mqttConnections/components/mqtt-connection-list/mqtt-connection-list.ts
@@ -10,6 +10,8 @@ import { CdkTableModule } from '@angular/cdk/table';
 import {MatTooltip} from '@angular/material/tooltip';
 import {MatCard, MatCardContent} from '@angular/material/card';
 import {MatSnackBar} from '@angular/material/snack-bar';
+import {MatDialog} from '@angular/material/dialog';
+import {ConfirmDialog, ConfirmDialogData} from '../../../../core/shared/components/confirm-dialog/confirm-dialog';
 
 @Component({
   selector: 'app-mqtt-connection-list',
@@ -39,6 +41,7 @@ export class MqttConnectionList implements OnInit {
   private mqttConnectionService = inject(MqttConnectionService);
   private router = inject(Router);
   private snack = inject(MatSnackBar);
+  private dialog = inject(MatDialog);
 
   mqttConnections = signal<MqttConnection[]>([]);
   loading = signal(false);
@@ -67,9 +70,23 @@ export class MqttConnectionList implements OnInit {
   }
 
   delete(id: number): void {
-    this.mqttConnectionService.deleteMqttConnection(id).subscribe({
-      next: () => {
-        this.load();
+    const dialog = this.dialog.open(ConfirmDialog, {
+      width: '400px',
+      data: {
+        title: 'Delete Connection',
+        message: 'Are you sure you want to delete this connection? <br><br>' +
+          'This also deletes every device, topic and rule associated with this mqtt connection. <br><br>' +
+          'This action cannot be undone.',
+        confirmText: 'Delete',
+        cancelText: 'Cancel'
+      } satisfies ConfirmDialogData
+    });
+
+    dialog.afterClosed().subscribe(confirmed => {
+      if (confirmed) {
+        this.mqttConnectionService.deleteMqttConnection(id).subscribe({
+          next: () => this.load()
+        });
       }
     });
   }
