+++ frontend/src/app/features/rules/components/rule-list/rule-list.ts
@@ -1,4 +1,4 @@
-import {Component, inject, signal} from '@angular/core';
+import {Component, inject, OnInit, signal} from '@angular/core';
 import {RuleService} from '../../../../core/services/rule.service';
 import {Router, RouterLink} from '@angular/router';
 import {Rule} from '../../../../core/models/rule.model';
@@ -17,6 +17,8 @@ import {MatIcon} from '@angular/material/icon';
 import {MatProgressSpinner} from '@angular/material/progress-spinner';
 import {MatTooltip} from '@angular/material/tooltip';
 import {RefactorDatePipe} from '../../../../core/pipes/refactorDate.pipe';
+import {MatDialog} from '@angular/material/dialog';
+import {ConfirmDialog, ConfirmDialogData} from '../../../../core/shared/components/confirm-dialog/confirm-dialog';
 
 @Component({
   selector: 'app-rule-list',
@@ -45,9 +47,10 @@ import {RefactorDatePipe} from '../../../../core/pipes/refactorDate.pipe';
   styleUrl: './rule-list.css',
 })
 
-export class RuleList {
+export class RuleList implements OnInit {
   private ruleService = inject(RuleService);
   private router = inject(Router);
+  private dialog = inject(MatDialog);
 
   rules = signal<Rule[]>([]);
   loading = signal(false);
@@ -76,9 +79,22 @@ export class RuleList {
   }
 
   delete(id: number): void {
-    this.ruleService.deleteRule(id).subscribe({
-      next: rules => {
-        this.load();
+    const dialog = this.dialog.open(ConfirmDialog, {
+      width: '400px',
+      data: {
+        title: 'Delete Rule',
+        message: 'Are you sure you want to delete this rule? <br><br>' +
+          'This action cannot be undone.',
+        confirmText: 'Delete',
+        cancelText: 'Cancel'
+      } satisfies ConfirmDialogData
+    });
+
+    dialog.afterClosed().subscribe(confirmed => {
+      if (confirmed) {
+        this.ruleService.deleteRule(id).subscribe({
+          next: () => this.load()
+        });
       }
     });
   }
