+++ frontend/src/app/features/topics/components/topic-list/topic-list.ts
@@ -16,6 +16,8 @@ import {
 import {MatIcon} from '@angular/material/icon';
 import {MatProgressSpinner} from '@angular/material/progress-spinner';
 import {MatTooltip} from '@angular/material/tooltip';
+import {MatDialog} from '@angular/material/dialog';
+import {ConfirmDialog, ConfirmDialogData} from '../../../../core/shared/components/confirm-dialog/confirm-dialog';
 
 @Component({
   selector: 'app-topic-list',
@@ -45,6 +47,7 @@ import {MatTooltip} from '@angular/material/tooltip';
 export class TopicList {
   private topicService = inject(TopicService);
   private router = inject(Router);
+  private dialog = inject(MatDialog);
 
   topics = signal<Topic[]>([]);
   loading = signal(false);
@@ -73,9 +76,23 @@ export class TopicList {
   }
 
   delete(id: number): void {
-    this.topicService.deleteTopic(id).subscribe({
-      next: topics => {
-        this.load();
+    const dialog = this.dialog.open(ConfirmDialog, {
+      width: '400px',
+      data: {
+        title: 'Delete Topic',
+        message: 'Are you sure you want to delete this rule? <br><br>' +
+          'This also deletes every rule associated with this topic. <br><br>' +
+          'This action cannot be undone.',
+        confirmText: 'Delete',
+        cancelText: 'Cancel'
+      } satisfies ConfirmDialogData
+    });
+
+    dialog.afterClosed().subscribe(confirmed => {
+      if (confirmed) {
+        this.topicService.deleteTopic(id).subscribe({
+          next: () => this.load()
+        });
       }
     });
   }
